requirements.txt
# Core
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Auth / Cors
django-cors-headers==4.3.1

# Filtering & pagination
django-filter==23.5

# Env & utils
python-dotenv==1.0.0
python-slugify==8.0.1
Pillow==10.1.0
requests==2.31.0

# Deploy
gunicorn==21.2.0
whitenoise==6.6.0
